requirements.txt
streamlit
pandas
pytesseract
Pillow
opencv-python-headless
pylibdmtx

